fix reachability for extensions
