+++ wire-ios-notification-engine/Sources/LegacyNotificationSessionLoader.swift
@@ -263,15 +263,12 @@ public struct LegacyNotificationSessionLoader {
         guard legacyCookieStorage.hasAuthenticationCookie else {
             throw Failure.mainAppRequired(message: "no authentication cookie")
         }
-        let reachabilityGroup = ZMSDispatchGroup(dispatchGroup: DispatchGroup(), label: "Sharing session reachability")
-        let serverNames = [legacyEnvironment.backendURL, legacyEnvironment.backendWSURL].compactMap(\.host)
-        let reachability = ZMReachability(serverNames: serverNames, group: reachabilityGroup)
         let transportSession = ZMTransportSession(
             environment: legacyEnvironment,
             proxyUsername: proxyCredentials?.username,
             proxyPassword: proxyCredentials?.password,
             cookieStorage: legacyCookieStorage,
-            reachability: reachability,
+            reachability: legacyEnvironment.reachability,
             initialAccessToken: nil,
             applicationGroupIdentifier: appGroupID,
             applicationVersion: buildNumber,
