+++ wire-ios-notification-engine/Sources/NotificationSession.swift
@@ -162,9 +162,6 @@ public final class NotificationSession {
             userIdentifier: accountIdentifier,
             useCache: false
         )
-        let reachabilityGroup = ZMSDispatchGroup(dispatchGroup: DispatchGroup(), label: "Sharing session reachability")
-        let serverNames = [environment.backendURL, environment.backendWSURL].compactMap(\.host)
-        let reachability = ZMReachability(serverNames: serverNames, group: reachabilityGroup)
 
         let credentials = environment.proxy.flatMap { ProxyCredentials.retrieve(for: $0) }
 
@@ -177,7 +174,7 @@ public final class NotificationSession {
             proxyUsername: credentials?.username,
             proxyPassword: credentials?.password,
             cookieStorage: cookieStorage,
-            reachability: reachability,
+            reachability: environment.reachability,
             initialAccessToken: nil,
             applicationGroupIdentifier: applicationGroupIdentifier,
             applicationVersion: "1.0.0",
