+++ wire-ios-sync-engine/Source/SessionManager/BackendEnvironmentProvider+Reachability.swift
@@ -23,19 +23,6 @@ public typealias Reachability = ReachabilityProvider & TearDownCapable
 
 extension BackendEnvironmentProvider {
 
-    public var reachability: ZMReachability {
-
-        let group = ZMSDispatchGroup(dispatchGroup: DispatchGroup(), label: "Reachability")
-
-        let serverNames: [String] = if let proxy {
-            [proxy.host]
-        } else {
-            [backendURL, backendWSURL].compactMap(\.host)
-        }
-
-        return ZMReachability(serverNames: serverNames, group: group)
-    }
-
     func reachabilityWrapper(enabled: Bool = false) -> ReachabilityWrapper {
         ReachabilityWrapper(enabled: enabled, reachabilityClosure: {
             self.reachability
